asdfa;sdlkja

+++ web/src/components/EditDeleteReportButtons.tsx
@@ -1,9 +1,19 @@
-import React from "react";
-import { Box, IconButton, Link } from "@chakra-ui/react";
+import React, { useState, useRef } from "react";
+import {
+	AlertDialog,
+	AlertDialogBody,
+	AlertDialogContent,
+	AlertDialogFooter,
+	AlertDialogHeader,
+	AlertDialogOverlay,
+	Box,
+	Button,
+	IconButton,
+	Link,
+} from "@chakra-ui/react";
 import NextLink from "next/link";
 import {
 	useDeleteIncidentReportMutation,
-	useDeletePostMutation,
 	useMeQuery,
 } from "../generated/graphql";
 import { DeleteIcon, EditIcon } from "@chakra-ui/icons";
@@ -18,6 +28,7 @@ export const EditDeleteReportButtons: React.FC<EditDeleteReportButtonsProps> = (
 	id,
 	creatorId,
 }) => {
+	const [isOpen, setIsOpen] = useState(false);
 	const router = useRouter();
 	const { data: meData } = useMeQuery();
 	const [deleteIncidentReport] = useDeleteIncidentReportMutation();
@@ -28,27 +39,55 @@ export const EditDeleteReportButtons: React.FC<EditDeleteReportButtonsProps> = (
 
 	return (
 		<Box>
-			<NextLink href="/post/edit/[id]" as={`/post/edit/${id}`}>
+			<NextLink
+				href="/incident-report/edit/[id]"
+				as={`/incident-report/edit/${id}`}
+			>
 				<IconButton
 					as={Link}
 					mr={4}
 					icon={<EditIcon />}
-					aria-label="Edit Post"
+					aria-label="Edit Report"
 				/>
 			</NextLink>
+
 			<IconButton
 				icon={<DeleteIcon />}
 				aria-label="Delete Post"
-				onClick={() => {
-					deleteIncidentReport({
-						variables: { id },
-						update: (cache) => {
-							cache.evict({ id: "IncidentReport:" + id });
-						},
-					});
-					router.push("/incidentReports");
-				}}
+				onClick={() => setIsOpen(true)}
 			/>
+			<AlertDialog isOpen={isOpen} onClose={() => setIsOpen(false)}>
+				<AlertDialogOverlay>
+					<AlertDialogContent>
+						<AlertDialogHeader fontSize="lg" fontWeight="bold">
+							Delete Incident Report
+						</AlertDialogHeader>
+
+						<AlertDialogBody>
+							Are you sure? You can't undo this action afterwards.
+						</AlertDialogBody>
+
+						<AlertDialogFooter>
+							<Button onClick={() => setIsOpen(false)}>Cancel</Button>
+							<Button
+								colorScheme="red"
+								onClick={() => {
+									deleteIncidentReport({
+										variables: { id },
+										update: (cache) => {
+											cache.evict({ id: "IncidentReport:" + id });
+										},
+									});
+									router.push("/incidentReports");
+								}}
+								ml={3}
+							>
+								Delete
+							</Button>
+						</AlertDialogFooter>
+					</AlertDialogContent>
+				</AlertDialogOverlay>
+			</AlertDialog>
 		</Box>
 	);
 };
